requirements.txt
Django>=4.0,<4.1.0
djangorestframework>=3.12<3.14.0
psycopg2>=2.9.1,<2.9.2
drf-spectacular>=0.15.1,<0.16
Pillow>=9.0.1,<9.2.0
uwsgi>=2.0.19,<2.1
requests>=2.26.0<2.28.0
Faker>=14.2.0<15.1.0
factory-boy==3.2.1
django-crontab>=0.7.0<0.7.1
APScheduler==3.6.3
uwsgi>=2.0.19,<2.1